+++ src/app/api/external/video-generation/route.ts
@@ -1,4 +1,3 @@
-import { NextResponse } from "next/server";
 import { videoGenerationSchema } from "@/features/video-generation/model/video-generation-schema";
 import { createMockVideoGeneration } from "@/server/video-generation/video-generation-store";
 import { requireApiKey } from "@/server/auth/api-key-guard";
@@ -7,54 +6,21 @@ import {
   buildGenerationSuccessResponse,
   buildInvalidRequestResponse,
 } from "@/server/http/response";
+import {
+  FILE_TOO_LARGE_ERROR,
+  getNumber,
+  getOptionalDataUrl,
+  getString,
+} from "@/server/http/form-data-utils";
 
 export const dynamic = "force-dynamic";
 export const revalidate = 0;
 export const runtime = "nodejs";
 
 const MAX_UPLOAD_BYTES = 10 * 1024 * 1024;
-const FILE_TOO_LARGE_ERROR = "FILE_TOO_LARGE";
-
-async function fileToDataUrl(file: File) {
-  const buffer = Buffer.from(await file.arrayBuffer());
-  const mime = file.type || "image/png";
-  return `data:${mime};base64,${buffer.toString("base64")}`;
-}
-
-function getString(formData: FormData, key: string) {
-  const value = formData.get(key);
-  if (typeof value !== "string") return undefined;
-  const trimmed = value.trim();
-  return trimmed.length > 0 ? trimmed : "";
-}
-
-function getNumber(formData: FormData, key: string) {
-  const value = getString(formData, key);
-  if (value === undefined) return undefined;
-  if (value === "") return NaN;
-  const parsed = Number(value);
-  return parsed;
-}
-
-async function getInitImageValue(formData: FormData) {
-  const entry = formData.get("initImage");
-  if (!entry) return undefined;
-  if (typeof entry === "string") {
-    return entry.trim() ? entry.trim() : "";
-  }
-  if (entry instanceof File) {
-    if (entry.size === 0) return undefined;
-    if (entry.size > MAX_UPLOAD_BYTES) {
-      throw new Error(FILE_TOO_LARGE_ERROR);
-    }
-    return fileToDataUrl(entry);
-  }
-  return undefined;
-}
-
 export async function POST(request: Request) {
   const auth = await requireApiKey(request);
-  if (auth instanceof NextResponse) {
+  if (auth instanceof Response) {
     return auth;
   }
 
@@ -65,7 +31,11 @@ export async function POST(request: Request) {
 
   let initImage: string | undefined;
   try {
-    initImage = await getInitImageValue(formData);
+    initImage = await getOptionalDataUrl(
+      formData,
+      "initImage",
+      MAX_UPLOAD_BYTES,
+    );
   } catch (error) {
     if (error instanceof Error && error.message === FILE_TOO_LARGE_ERROR) {
       return buildErrorResponse("FILE_TOO_LARGE", 413);
